requirements.txt
django>=5.0.0
djangorestframework
pyyaml
djangorestframework-simplejwt
pillow
django-filter
#mysql
#mysql-connector-python
#mysql-connector-python
#install latest version of node must be > 18 

